requirements.txt
googletrans
openai-whisper
moviepy
whisperx